% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorse.R
\name{get_data}
\alias{get_data}
\title{Get indicator data from the NBS website.}
\usage{
get_data(
  indicators,
  period = "1949-",
  database = c("fsnd", "hgnd", "fsjd", "hgjd", "fsyd", "hgyd"),
  language = c("en", "zh"),
  nap_control = nap(~. > 20L, runif, 1L, 0.5, 1)
)
}
\arguments{
\item{indicators}{a character vector of indicator codes like \code{"A010101"}
or \code{c("A010101", "A010102")}.}

\item{period}{the period of the time series. Below shows a list of valid
period filters:
\itemize{
\item{month: }{201201, 201205.}
\item{quarter: }{2012A, 2012B, 2012C, 2012D.}
\item{year: }{2012, 2013.}
\item{duration: }{2013-, 2008-2009, last10.}
}}

\item{database}{see the \code{database} argument in \code{\link{get_meta}}.}

\item{language}{either "en" (i.e. English) or "zh" (i.e. Chinese).}

\item{nap_control}{let the function take a nap in-between two query calls. No
nap to be taken for a short list of indicators (less than 20); a nap of
length defaulting to be a random number between 0.5s and 1s is invoked for
each indicator in addition to the first 20. See \code{\link{nap}} for more
details on how to control the pause. This makes the querier less likely to
be flagged by the anti-bot mechanism on the NBS website.}
}
\value{
A list of two dataframes: indicators and their metadata.
}
\description{
Retrieve indicators sequentially from a database.
}
\examples{
\dontrun{
 get_data("A010101")
 get_data("A010101", database = "fsnd")
}
}
