% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorse.R
\name{nap}
\alias{nap}
\title{Control the naps in-between query calls.}
\usage{
nap(when = ~TRUE, how = runif, ...)
}
\arguments{
\item{when}{a function that takes an integer as the only argument and returns
a single \code{TRUE/FALSE}; it could be a purrr-style lambda. See
\code{\link[purrr]{as_mapper}} for more details on the syntax. That integer
represents the position of the current query. For example, \code{~.x > 10}
means the querier only takes naps after the first 10 calls. Default to be
\code{~TRUE}, i.e., always take a nap in-between two queries.}

\item{how}{a function that generates a single random number to be used as the
nap length. Default to be \code{\link[stats]{runif}}.}

\item{...}{arguments to be passed into the \code{how} function.}
}
\value{
a function that generates a single random number conditional upon
\code{when} and \code{how}.
}
\description{
An auxiliary function to control the pauses in-between query
calls so as to avoid being flagged by the anti-bot mechanism on the NBS
website.
}
\examples{
\dontrun{
# always take a nap of 0.5-1s in-between two queries.
nap(~TRUE, runif, 1L, 0.5, 1)
# never take naps for the first 20 calls; take naps of 0.5-1s for the rest.
nap(~. > 20L, runif, 1L, 0.5, 1)
}
}
